[refactor] : file structure formatted - moved some components from src/components to src/features/Home - added default exports in index.js file in Home dir
