+++ src/features/Home/index.js
@@ -1,0 +1,5 @@
+export { default as WeeklyWorkoutReport } from './WeeklyWorkoutReport';
+export { default as FitnessScore } from './FitnessScore';
+export { default as LeaderBoard } from './LeaderBoard';
+export { default as DuePaymentIndicator } from './DuePaymentIndicator';
+export { default as MoveCoins } from './MoveCoins';
